==> 3_Task2_data_transformation.ipynb <==
# Task 2: Predicting customer buying behaviour
## 3. Data transformation

## Import libraries

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.preprocessing import OneHotEncoder
from sklearn.preprocessing import OrdinalEncoder
from category_encoders import TargetEncoder

from sklearn.preprocessing import QuantileTransformer

from sklearn.preprocessing import MinMaxScaler

## Load the data

cat = pd.read_pickle("cat_eda.pickle")
num = pd.read_pickle("num_eda.pickle")

## Separate the target variable

target = num[["booking_complete"]].copy().reset_index(drop=True)

## CATEGORICAL VARIABLES TRANSFORMATION

cat.head()

### One Hot Encoding

#### Variables to apply OHE

var_ohe = ["sales_channel","trip_type","route","booking_origin"]

#### Instantiate

ohe = OneHotEncoder(sparse_output = False, handle_unknown='ignore')

#### Train and apply

cat_ohe = ohe.fit_transform(cat[var_ohe])

#### Save as dataframe

cat_ohe = pd.DataFrame(cat_ohe, columns = ohe.get_feature_names_out())
cat_ohe.head()

### Ordinal Encoding

#### Variables to apply OE

var_oe = ["flight_day"]

#### Order of those variables

order_flight_day = ["Mon","Tue","Wed","Thu","Fri","Sat","Sun"]

#### Instantiate

oe = OrdinalEncoder(categories = [order_flight_day],
                    handle_unknown = 'use_encoded_value',
                    unknown_value = 10)

#### Fit and transform

cat_oe = oe.fit_transform(cat[var_oe]) + 1

#### Save as a dataframe

#Add suffixes to variables' names
names_oe = [variable + '_oe' for variable in var_oe]

#Save as a dataframe
cat_oe = pd.DataFrame(cat_oe, columns = names_oe)
cat_oe.head()

### Target Encoding

#### Variables to apply TE

# var_te = ["route","booking_origin"]

#### Instantiate

# te = TargetEncoder(min_samples_leaf=100, return_df = False)

#### Fit and transform

# cat_te = te.fit_transform(cat[var_te].reset_index(drop=True), y = target)

#### Save as a dataframe

# #Add suffixes to variables' names
# names_te = [variable + '_te' for variable in var_te]

# #Save as a dataframe
# cat_te = pd.DataFrame(cat_te, columns = names_te)
# cat_te.head()

## NUMERICAL VARIABLES TRANSFORMATION

### Normalization (Gauss)

var_qt = ["num_passengers","purchase_lead","length_of_stay","flight_hour","flight_duration"]

##### Instantiate

qt = QuantileTransformer(output_distribution='normal')

##### Fit and transform

num_qt = qt.fit_transform(num[var_qt])

##### Save as a dataframe

#Add suffixes to variables' names
names_qt = [variable + '_qt' for variable in var_qt]

#Save as a dataframe
num_qt = pd.DataFrame(num_qt,columns = names_qt)

## COMBINE ALL DATASETS

First let's add the non-transformed data

variables = cat.columns.to_list() + num.iloc[:,:-1].columns.to_list()
var_trans = var_oe + var_ohe + var_qt
var_nt = [name for name in variables if name not in var_trans]
var_nt

num_nt = num[var_nt].reset_index().iloc[:,1:]

dataframes = []
dataframes.extend(value for name, value in locals().items() if name.startswith('cat_') or name.startswith('num_'))

### Concatenate all the datasets

df = pd.concat(dataframes, axis = 1)
df

## FEATURE SCALING

### Min-Max scaling (Normalization between 0 and 1)

df.info()

var_mms = df.iloc[:,73:78].columns
var_mms

#### Instantiate

mms = MinMaxScaler()

#### Fit and transform

df_mms = mms.fit_transform(df[var_mms])

#### Save as a dataframe

#Add suffixes to variables' names
nombres_mms = [variable + '_mms' for variable in var_mms]

#Save as a dataframe
df_mms = pd.DataFrame(df_mms,columns = nombres_mms)

### COMBINE SCALED DATASETS

include = [df, df_mms, target]

df_final = pd.concat(include, axis=1)
df_final.head()

Save the final dataframe in .pickle format to proceed to the next step (modeling).

pd.to_pickle(df_final, "df_transformed.pickle")
