==> 5_Task2_data_balancing.ipynb <==
# Task 2: Predicting customer buying behaviour
## 5. Data balancing

## Import libraries

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

from imblearn.under_sampling import RandomUnderSampler
from imblearn.over_sampling import RandomOverSampler
from imblearn.under_sampling import TomekLinks
from imblearn.over_sampling import SMOTE
from imblearn.combine import SMOTETomek

from sklearn.model_selection import train_test_split

from sklearn.linear_model import LogisticRegression

from sklearn.metrics import roc_auc_score

## Load the data

x = pd.read_pickle("x_final.pickle")
y = pd.read_pickle("y_final.pickle")

x.head()

## Data balancing

The main goal is to evaluate if balancing the data increases the prediction of the dataset.

To do that different methods should be tested.

### No balancing

#### Split between train and test

train_x,test_x,train_y,test_y = train_test_split(x,y,test_size=0.3)

#### Instantiate the model (a simple logistic regression model)

rl_no_balancing = LogisticRegression(n_jobs = -1)

#### Fit the model on the train set

rl_no_balancing.fit(train_x,train_y)

#### Predict probabilities

pred_rl_no_balancing = rl_no_balancing.predict_proba(test_x)[:,1]

#### Evaluate model

roc_rl_no_balancing = roc_auc_score(test_y, pred_rl_no_balancing)

roc_rl_no_balancing

### Balance the dataset using undersampling

#### Instantiate the undersampler

rus = RandomUnderSampler(sampling_strategy= 1)

#### Fit and "transform" the undersampler

x_rus, y_rus = rus.fit_resample(x,y)

#### Split train and test

train_x_rus,test_x_rus,train_y_rus,test_y_rus = train_test_split(x_rus,y_rus,test_size=0.3)

#### Instantiate the model

rl_rus = LogisticRegression(n_jobs = -1)

#### Fit the model on the train set

rl_rus.fit(train_x_rus,train_y_rus)

#### Predict probabilities

pred_rl_rus = rl_rus.predict_proba(test_x)[:,1]

#### Evaluate model

roc_rl_rus = roc_auc_score(test_y, pred_rl_rus)

roc_rl_rus

### Balance the dataset using oversampling

ros = RandomOverSampler(sampling_strategy= 1)

#### Instantiate and "transform" the oversampler

x_ros, y_ros = ros.fit_resample(x,y)

#### Split train and test

train_x_ros,test_x_ros,train_y_ros,test_y_ros = train_test_split(x_ros,y_ros,test_size=0.3)

#### Instantiate the model

rl_ros = LogisticRegression(n_jobs = -1)

#### Fit the model on the train set

rl_ros.fit(train_x_ros,train_y_ros)

#### Predict probabilities

pred_rl_ros = rl_ros.predict_proba(test_x)[:,1]

#### Evaluate model

roc_rl_ros = roc_auc_score(test_y, pred_rl_ros)

roc_rl_ros

#### Since there is no significant difference using undersampling or oversampling, the data is not going to be balanced.

Save the final dataframe in .pickle format to proceed to the next step (modeling).

pd.to_pickle(x, "x_bal.pickle")
pd.to_pickle(y, "y_bal.pickle")
