==> 7_Task2_preproduction.ipynb <==
# Task 2: Predicting customer buying behaviour
## 7. Preproduction code

## Import libraries

import numpy as np
import pandas as pd
import cloudpickle

from sklearn.preprocessing import OneHotEncoder
from sklearn.preprocessing import OrdinalEncoder

from sklearn.preprocessing import QuantileTransformer

from sklearn.preprocessing import MinMaxScaler

from xgboost import XGBClassifier

from sklearn.pipeline import Pipeline
from sklearn.preprocessing import FunctionTransformer
from sklearn.compose import make_column_transformer
from sklearn.pipeline import make_pipeline

## Load the data (original dataset)

df = pd.read_csv("customer_booking.csv",encoding='ISO-8859-1')
df

## Enumerate final variables from feature selection

variables = ["booking_origin","length_of_stay","purchase_lead","flight_duration","flight_hour","flight_day","num_passengers","sales_channel","wants_extra_baggage","wants_preferred_seat","wants_in_flight_meals","route"]
variables

## Modify datasets before proceeding to building the pipeline

### Remove duplicates

df.drop_duplicates(inplace = True)

### Select final variables from df

x = df[variables].reset_index(drop=True).copy()

target = "booking_complete"
y = df[target].reset_index(drop=True).copy()

## Build pipeline

### Data cleaning summary

#### Build function

def data_cleaning(df):
    # Execute all the steps carried out in step 1 (data cleaning script)
    temp = df
    # Removing outliers from purchase_lead and length_of_stay
    var_clip = ["purchase_lead","length_of_stay"]
    p_min = 0
    p_max = 0.99

    for variable in var_clip:
        temp[variable].clip(lower = temp[variable].quantile(p_min), 
                        upper = temp[variable].quantile(p_max), 
                        inplace = True)
        
    return temp

#### Convert function to transformer

So we can add it to the pipeline

trans_data_cleaning = FunctionTransformer(data_cleaning)

### Feature engineering summary

var_ohe = ["sales_channel","route","booking_origin"]
ohe = OneHotEncoder(sparse_output = False, handle_unknown='ignore')

var_oe = ["flight_day"]
order_flight_day = ["Mon","Tue","Wed","Thu","Fri","Sat","Sun"]
oe = OrdinalEncoder(categories = [order_flight_day],
                    handle_unknown = 'use_encoded_value',
                    unknown_value = 10)

var_qt = ["num_passengers","purchase_lead","length_of_stay","flight_hour","flight_duration"]
qt = QuantileTransformer(output_distribution='normal')

var_mms = ["num_passengers","purchase_lead","length_of_stay","flight_hour"]
mms = MinMaxScaler()

### Build preprocessing pipe

#### Create column transformer

ct = make_column_transformer(
    (ohe, var_ohe),
    (oe, var_oe),
    (qt, var_qt),
    (mms, var_mms),
    remainder='drop')

### Build preprocessing pipeline

pipe_prepro = make_pipeline(trans_data_cleaning, ct)

### Instantiate the model

model = XGBClassifier(n_jobs=-1, max_depth=5, learning_rate = 0.1, n_estimators=100)

### Create training pipeline

pipe_training = make_pipeline(pipe_prepro, model)

### Save the training pipeline

with open("pipe_training.pickle",mode="wb") as file:
    cloudpickle.dump(pipe_training, file)

### Train the pipeline and create production pipeline

pipe_production = pipe_training.fit(x,y)

## Save execution pipeline

with open("pipe_production.pickle",mode="wb") as file:
    cloudpickle.dump(pipe_production, file)
