==> 8_Task2_retraining.ipynb <==
# Task 2: Predicting customer buying behaviour
## 8. Retraining code

#### This script should be run some time after deployment of the production code. It is used to retrain the machine learning model once its accuracy has decreased.

import numpy as np
import pandas as pd
import cloudpickle


df = pd.read_csv("customer_booking.csv",encoding='ISO-8859-1')

df.drop_duplicates(inplace=True)

variables = ["booking_origin","length_of_stay","purchase_lead","flight_duration","flight_hour","flight_day","num_passengers","sales_channel","wants_extra_baggage","wants_preferred_seat","wants_in_flight_meals","route"]

target = "booking_complete"

x = df[variables].copy()
y = df[target].copy()

# Open pipe_training
with open("pipe_training.pickle", mode='rb') as file:
   pipe_training = cloudpickle.load(file)

# Fit the training pipe and save the new production pipeline
pipe_production = pipe_training.fit(x,y)

with open("pipe_production.pickle", mode='wb') as file:
   cloudpickle.dump(pipe_production, file)
