==> 9_Task2_production.ipynb <==
# Task 2: Predicting customer buying behaviour
## 9. Production code

#### This is the script that shoud be run in .py format in production. It can be called from an API for real-time applications for example.

import numpy as np
import pandas as pd
import cloudpickle

# The data read as df should the NEW data. 
# In this case it could have been a random sample from the original df.
df = pd.read_csv("customer_booking.csv",encoding='ISO-8859-1')

df.drop_duplicates(inplace=True)

variables = ["booking_origin","length_of_stay","purchase_lead","flight_duration",
                "flight_hour","flight_day","num_passengers","sales_channel",
                "wants_extra_baggage","wants_preferred_seat","wants_in_flight_meals",
                "route"]

df = df[variables]

# Open pipe_production
with open("pipe_production.pickle", mode='rb') as file:
   pipe_production = cloudpickle.load(file)

scoring = pipe_production.predict_proba(df)[:, 1]

### The scoring returns the probability of a customer booking a flight with British Airways. 

### In order to interpret the result, different discrimination thresholds can be used in order to assess further steps into the booking process of a customer.

For example, requiring a high discrimination threshold (let's say 80%) will lead to a small number of predicted actual bookings and thus increasing the precision of the model. However, many actual bookings would be predicted as non-bookings and so the model would not capture them (low recall). This would reduce the cost for the company in case they wanted to provide further action in the booking process (e.g. discounts for the customers) but many booking opportunities would not be considered (cost of false negatives).

On the other hand, requiring a low discrimination threshold (let's say 10%) would lead to a large number of false positives. That is, many predictions as true bookings but many of them are wrong (low precision and high recall). This would increase the cost of false positives but probably having more customer booking

### Then, further steps should be carried out in order to appropriately interpret the results from the machine learning model from a business perspective.
